Move Logger class to separate package
